Refactor: rename error handler utility module
