+++ lua/spring-initializr/utils/error_handler_utils.lua
